refactor: separado os testes de cada classe em arquivos diferentes e iniciado os testes de reserva
