--- Testes/UsuarioTest.java
@@ -1,20 +1,0 @@
-import static org.junit.jupiter.api.Assertions.*;
-
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-
-class UsuarioTest {
-
-	@BeforeAll
-	void before() {
-		InterfaceUsuario interf = new InterfaceUsuario(); 
-		
-	}
-	
-	@Test
-	void CadastrarUsuarioTest() {
-		
-	}
-	
-	
-}
